+++ halotools/mock_observables/pair_counters/__init__.py
@@ -11,8 +11,5 @@ from .marked_npairs_xy_z import marked_npairs_xy_z
 from .npairs_jackknife_3d import npairs_jackknife_3d
 from .npairs_s_mu import npairs_s_mu
 from .npairs_per_object_3d import npairs_per_object_3d
-from .velocity_marked_npairs_3d import velocity_marked_npairs_3d
 from .pairwise_distance_3d import pairwise_distance_3d
 from .pairwise_distance_xy_z import pairwise_distance_xy_z
-from .velocity_marked_npairs_3d import velocity_marked_npairs_3d
-from .velocity_marked_npairs_xy_z import velocity_marked_npairs_xy_z
