+++ halotools/mock_observables/pair_counters/marked_cpairs/__init__.py
@@ -5,5 +5,3 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from .conditional_pairwise_distances import *
 from .marked_npairs_3d_engine import marked_npairs_3d_engine
 from .marked_npairs_xy_z_engine import marked_npairs_xy_z_engine
-from .velocity_marked_npairs_3d_engine import velocity_marked_npairs_3d_engine
-from .velocity_marked_npairs_xy_z_engine import velocity_marked_npairs_xy_z_engine
