+++ halotools/mock_observables/pairwise_velocities/__init__.py
@@ -10,3 +10,5 @@ from .mean_radial_velocity_vs_r import mean_radial_velocity_vs_r
 from .radial_pvd_vs_r import radial_pvd_vs_r
 from .mean_los_velocity_vs_rp import mean_los_velocity_vs_rp
 from .los_pvd_vs_rp import los_pvd_vs_rp
+from .velocity_marked_npairs_3d import velocity_marked_npairs_3d
+from .velocity_marked_npairs_xy_z import velocity_marked_npairs_xy_z
