+++ halotools/mock_observables/pairwise_velocities/engines/__init__.py
@@ -1,0 +1,2 @@
+from .velocity_marked_npairs_3d_engine import velocity_marked_npairs_3d_engine
+from .velocity_marked_npairs_xy_z_engine import velocity_marked_npairs_xy_z_engine
