+++ halotools/mock_observables/pairwise_velocities/los_pvd_vs_rp.py
@@ -11,7 +11,7 @@ import numpy as np
 from .pairwise_velocities_helpers import (_pairwise_velocity_stats_process_args, 
     _process_rp_bins)
 
-from ..pair_counters import velocity_marked_npairs_xy_z
+from .velocity_marked_npairs_xy_z import velocity_marked_npairs_xy_z
 
 __all__ = ('los_pvd_vs_rp', )
 __author__ = ['Duncan Campbell']
