+++ halotools/mock_observables/pairwise_velocities/velocity_marked_npairs_xy_z.py
@@ -3,11 +3,11 @@ import numpy as np
 import multiprocessing
 from functools import partial 
 
-from .npairs_xy_z import _npairs_xy_z_process_args
-from .mesh_helpers import _set_approximate_cell_sizes, _cell1_parallelization_indices
-from .rectangular_mesh import RectangularDoubleMesh
+from ..pair_counters.npairs_xy_z import _npairs_xy_z_process_args
+from ..pair_counters.mesh_helpers import _set_approximate_cell_sizes, _cell1_parallelization_indices
+from ..pair_counters.rectangular_mesh import RectangularDoubleMesh
 from .velocity_marked_npairs_3d import _func_signature_int_from_vel_weight_func_id
-from .marked_cpairs import velocity_marked_npairs_xy_z_engine 
+from .engines import velocity_marked_npairs_xy_z_engine 
 
 from ...utils.array_utils import convert_to_ndarray
 from ...custom_exceptions import HalotoolsError
